moved media functions to mediaConfig.ts, fixed skeleton height
