--- aim/web/ui/src/utils/getAudioBoxSize.ts
@@ -1,8 +1,0 @@
-import {
-  AUDIO_FIXED_HEIGHT,
-  AUDIO_FIXED_WIDTH,
-} from 'config/mediaConfigs/mediaConfigs';
-
-export default function getAudioBoxSize(): { width: number; height: number } {
-  return { width: AUDIO_FIXED_WIDTH, height: AUDIO_FIXED_HEIGHT };
-}
