--- aim/web/ui/src/utils/getAudioMediaListHeight.ts
@@ -1,7 +1,0 @@
-import { ITEM_CAPTION_HEIGHT } from 'config/mediaConfigs/mediaConfigs';
-
-export default function getAudioMediaListHeight(
-  mediaItemHeight: number,
-): number {
-  return mediaItemHeight + ITEM_CAPTION_HEIGHT;
-}
